decouple AssetsManager from dex wallet constructors
