+++ libwallet/assets_manager.go
@@ -7,13 +7,10 @@ import (
 	"path/filepath"
 	"sort"
 	"strconv"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
-	"decred.org/dcrdex/client/asset"
-	"decred.org/dcrdex/dex"
 	"decred.org/dcrwallet/v3/errors"
 	"github.com/asdine/storm"
 	"github.com/asdine/storm/q"
@@ -26,22 +23,15 @@ import (
 	"github.com/crypto-power/cryptopower/ui/values"
 	bolt "go.etcd.io/bbolt"
 
-	dexbtc "decred.org/dcrdex/client/asset/btc"
-	dexDcr "decred.org/dcrdex/client/asset/dcr"
-	dexltc "decred.org/dcrdex/client/asset/ltc"
-	btccfg "github.com/btcsuite/btcd/chaincfg"
 	"github.com/crypto-power/cryptopower/libwallet/assets/btc"
 	"github.com/crypto-power/cryptopower/libwallet/assets/dcr"
 	"github.com/crypto-power/cryptopower/libwallet/assets/ltc"
 	sharedW "github.com/crypto-power/cryptopower/libwallet/assets/wallet"
-	dcrcfg "github.com/decred/dcrd/chaincfg/v3"
 )
 
 // TODO: This is the main app's log filename, should probably be defined
 // elsewhere.
 const LogFilename = "cryptopower.log"
-
-var dexWalletRegistered atomic.Bool
 
 // Assets is a struct that holds all the assets supported by the wallet.
 type Assets struct {
@@ -926,12 +916,6 @@ func (mgr *AssetsManager) InitializeDEX(ctx context.Context) {
 		return
 	}
 
-	if !dexWalletRegistered.Load() {
-		mgr.prepareDexSupportForDCRWallet()
-		mgr.prepareDexSupportForBTCCloneWallets()
-		dexWalletRegistered.Store(true)
-	}
-
 	if mgr.DexcInitialized() || mgr.startingDEX.Load() {
 		log.Debug("Attempted to reinitialize a running dex client instance")
 		return
@@ -942,6 +926,11 @@ func (mgr *AssetsManager) InitializeDEX(ctx context.Context) {
 	defer func() {
 		mgr.startingDEX.Store(false)
 	}()
+
+	// Initialize support for accessing Cryptopower wallets from DEX before
+	// starting the DEX client so the DEX client can bind previously added
+	// wallets when it starts.
+	setDEXWalletLoader(mgr.WalletWithID)
 
 	logDir := filepath.Dir(mgr.LogFile())
 	dexcl, err := dexc.Start(ctx, mgr.RootDir(), mgr.GetLanguagePreference(), logDir, mgr.GetLogLevels(), mgr.NetType(), 0 /* TODO: Make configurable */)
@@ -980,160 +969,4 @@ func (mgr *AssetsManager) DeleteDEXData() error {
 
 	// Delete dex client db.
 	return os.Remove(dexDBFile)
-}
-
-// prepareDexSupportForDCRWallet sets up the DEX client to allow using a
-// cyptopower dcr wallet with DEX core.
-func (mgr *AssetsManager) prepareDexSupportForDCRWallet() {
-	// Build a custom wallet definition with custom config options
-	// for use by the dex dcr ExchangeWallet.
-	customWalletConfigOpts := []*asset.ConfigOption{
-		{
-			Key:         dexc.WalletIDConfigKey,
-			DisplayName: "Wallet ID",
-			Description: "ID of existing wallet to use",
-		},
-		{
-			Key:         dexc.WalletAccountNumberConfigKey,
-			DisplayName: "Wallet Account Number",
-			Description: "Account number of the selected wallet",
-		},
-	}
-
-	def := &asset.WalletDefinition{
-		Type:        dexc.CustomDexWalletType,
-		Description: "Uses an existing cryptopower Wallet.",
-		ConfigOpts:  customWalletConfigOpts,
-	}
-
-	// This function will be invoked when the DEX client needs to
-	// setup a dcr ExchangeWallet; it allows us to use an existing
-	// wallet instance for wallet operations instead of json-rpc.
-	var walletMaker = func(settings map[string]string, chainParams *dcrcfg.Params, logger dex.Logger) (dexDcr.Wallet, error) {
-		walletIDStr := settings[dexc.WalletIDConfigKey]
-		walletID, err := strconv.Atoi(walletIDStr)
-		if err != nil || walletID < 0 {
-			return nil, fmt.Errorf("invalid wallet ID %q in settings", walletIDStr)
-		}
-
-		wallet := mgr.WalletWithID(walletID)
-		if wallet == nil {
-			return nil, fmt.Errorf("no wallet exists with ID %q", walletIDStr)
-		}
-
-		walletParams := wallet.Internal().DCR.ChainParams()
-		if walletParams.Net != chainParams.Net {
-			return nil, fmt.Errorf("selected wallet is for %s network, expected %s", walletParams.Name, chainParams.Name)
-		}
-
-		if wallet.IsWatchingOnlyWallet() {
-			return nil, fmt.Errorf("cannot use watch only wallet for DEX trade")
-		}
-
-		// Ensure the account exists.
-		accountNumberStr := settings[dexc.WalletAccountNumberConfigKey]
-		acctNum, err := strconv.ParseInt(accountNumberStr, 10, 64)
-		if err != nil {
-			return nil, err
-		}
-
-		accountNumber := int32(acctNum)
-		if _, err = wallet.AccountName(accountNumber); err != nil {
-			return nil, fmt.Errorf("error checking selected DEX account: %w", err)
-		}
-
-		dcrAsset, ok := wallet.(*dcr.Asset)
-		if !ok {
-			return nil, fmt.Errorf("DEX wallet not supported for %s", walletParams.Name)
-		}
-
-		return dcr.NewDEXWallet(dcrAsset, accountNumber, dcrAsset.SyncData()), nil
-	}
-
-	dexDcr.RegisterCustomWallet(walletMaker, def)
-}
-
-// prepareDexSupportForBTCCloneWallets sets up the DEX client to allow using a
-// Cyptopower btc or ltc wallet with DEX core.
-func (mgr *AssetsManager) prepareDexSupportForBTCCloneWallets() {
-	// Build a custom wallet definition with custom config options for use by
-	// the dexbtc.ExchangeWalletSPV.
-	customWalletConfigOpts := []*asset.ConfigOption{
-		{
-			Key:         dexc.WalletIDConfigKey,
-			DisplayName: "Wallet ID",
-			Description: "ID of existing wallet to use",
-		},
-		{
-			Key:         dexc.WalletAccountNumberConfigKey,
-			DisplayName: "Wallet Account Number",
-			Description: "Account number of the selected wallet",
-		},
-	}
-
-	def := &asset.WalletDefinition{
-		Type:        dexc.CustomDexWalletType,
-		Description: "Uses an existing cryptopower Wallet.",
-		ConfigOpts:  customWalletConfigOpts,
-	}
-
-	// Register wallet constructors. The constructor function will be invoked
-	// when the DEX client needs to setup a dexbtc.BTCWallet and this allows us
-	// to use an existing wallet instance for wallet operations.
-
-	btcWalletConstructor := func(settings map[string]string, chainParams *btccfg.Params) (dexbtc.BTCWallet, error) {
-		return mgr.btcCloneWalletConstructor(false, settings, chainParams)
-	}
-	dexbtc.RegisterCustomSPVWallet(btcWalletConstructor, def)
-
-	ltcWalletConstructor := func(settings map[string]string, chainParams *btccfg.Params) (dexbtc.BTCWallet, error) {
-		return mgr.btcCloneWalletConstructor(true, settings, chainParams)
-	}
-	dexltc.RegisterCustomSPVWallet(ltcWalletConstructor, def)
-}
-
-// btcCloneWalletConstructor is a shared wallet constructor used by btc and ltc
-// to create dex compatible wallets.
-func (mgr *AssetsManager) btcCloneWalletConstructor(isLtc bool, settings map[string]string, chainParams *btccfg.Params) (dexbtc.BTCWallet, error) {
-	walletIDStr := settings[dexc.WalletIDConfigKey]
-	walletID, err := strconv.Atoi(walletIDStr)
-	if err != nil || walletID < 0 {
-		return nil, fmt.Errorf("invalid wallet ID %q in settings", walletIDStr)
-	}
-
-	wallet := mgr.WalletWithID(walletID)
-	if wallet == nil {
-		return nil, fmt.Errorf("no wallet exists with ID %q", walletIDStr)
-	}
-
-	if isLtc {
-		if walletParams := wallet.Internal().LTC.ChainParams(); !strings.EqualFold(walletParams.Name, chainParams.Name) {
-			return nil, fmt.Errorf("selected wallet is for %s network, expected %s", walletParams.Name, chainParams.Name)
-		}
-	} else {
-		if walletParams := wallet.Internal().BTC.ChainParams(); walletParams.Net != chainParams.Net {
-			return nil, fmt.Errorf("selected wallet is for %s network, expected %s", walletParams.Name, chainParams.Name)
-		}
-	}
-
-	if wallet.IsWatchingOnlyWallet() {
-		return nil, fmt.Errorf("cannot use watch only wallet for DEX trade")
-	}
-
-	// Ensure the wallet account exists.
-	accountNumberStr := settings[dexc.WalletAccountNumberConfigKey]
-	acctNum, err := strconv.ParseInt(accountNumberStr, 10, 64)
-	if err != nil {
-		return nil, err
-	}
-
-	accountNumber := int32(acctNum)
-	if _, err = wallet.AccountName(accountNumber); err != nil {
-		return nil, fmt.Errorf("error checking selected DEX account name: %w", err)
-	}
-
-	if isLtc {
-		return ltc.NewDEXWallet(wallet.Internal().LTC, accountNumber, wallet.(*ltc.Asset).NeutrinoClient(), chainParams), nil
-	}
-	return btc.NewDEXWallet(wallet.Internal().BTC, accountNumber, wallet.(*btc.Asset).NeutrinoClient()), nil
 }
